# Conflicts:
#	Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConversationCell.m

+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConversationCell.m
@@ -312,7 +312,11 @@ const NSTimeInterval ConversationCellSelectionAnimationDuration = 0.33;
     
     self.toolboxHeightConstraint.active = ! shouldBeVisible;
     self.likeButton.alpha = shouldBeVisible ? 1 : 0;
-
+    
+    [UIView performWithoutAnimation:^{
+        [self layoutIfNeeded];
+    }];
+    
     if (animated) {
         [UIView animateWithDuration:0.35 animations:^{
             self.messageToolboxView.alpha = shouldBeVisible ? 1 : 0;
@@ -431,7 +435,9 @@ const NSTimeInterval ConversationCellSelectionAnimationDuration = 0.33;
     UIMenuController *menuController = [UIMenuController sharedMenuController];
     UIMenuItem *deleteItem = [[UIMenuItem alloc] initWithTitle:NSLocalizedString(@"content.message.delete", @"") action:@selector(deleteMessage:)];
     NSArray <UIMenuItem *> *items = menuConfigurationProperties.additionalItems ?: @[];
-    menuController.menuItems = [items arrayByAddingObject:deleteItem];
+    if (self.message.deliveryState == ZMDeliveryStateDelivered) {
+        menuController.menuItems = [items arrayByAddingObject:deleteItem];
+    }
     [menuController setTargetRect:menuConfigurationProperties.targetRect inView:menuConfigurationProperties.targetView];
     [menuController setMenuVisible:YES animated:YES];
